Also fix VcpkgPaths.h forward declaring stuff that it does not own.
