+++ include/vcpkg/base/fwd/system.process.h
@@ -2,11 +2,6 @@
 
 namespace vcpkg
 {
-    struct CMakeVariable;
-    struct Command;
-    struct CommandLess;
-    struct ExitCodeAndOutput;
-
     enum class EchoInDebug
     {
         Show,
@@ -18,4 +13,11 @@ namespace vcpkg
         Utf8,
         Utf16
     };
+
+    struct CMakeVariable;
+    struct Command;
+    struct CommandLess;
+    struct ExitCodeAndOutput;
+    struct Environment;
+    struct WorkingDirectory;
 }
