+++ include/vcpkg/fwd/binaryparagraph.h
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace vcpkg
+{
+    struct BinaryParagraph;
+    struct BinaryControlFile;
+}
