+++ include/vcpkg/fwd/triplet.h
@@ -2,5 +2,6 @@
 
 namespace vcpkg
 {
+    struct TripletInstance;
     struct Triplet;
 }
