+++ include/vcpkg/fwd/vcpkgpaths.h
@@ -4,5 +4,9 @@ namespace vcpkg
 {
     struct ToolsetArchOption;
     struct Toolset;
+    struct ManifestAndPath;
+    struct TripletFile;
     struct VcpkgPaths;
+
+    struct VcpkgPathsImpl;
 }
