+++ include/vcpkg/vcpkgpaths.h
@@ -4,6 +4,8 @@
 #include <vcpkg/base/fwd/json.h>
 #include <vcpkg/base/fwd/system.process.h>
 
+#include <vcpkg/fwd/binaryparagraph.h>
+#include <vcpkg/fwd/build.h>
 #include <vcpkg/fwd/configuration.h>
 #include <vcpkg/fwd/installedpaths.h>
 #include <vcpkg/fwd/registries.h>
@@ -41,16 +43,6 @@ namespace vcpkg
         std::string full_version;
         std::vector<ToolsetArchOption> supported_architectures;
     };
-
-    struct PreBuildInfo;
-    struct AbiInfo;
-    struct CompilerInfo;
-    struct VcpkgPathsImpl;
-
-    struct BinaryParagraph;
-    struct Environment;
-    struct PackageSpec;
-    struct Triplet;
 
     struct ManifestAndPath
     {
